+++ tests/AZOA.WebAPI.IntegrationTests/Bridge/BridgeSafetyHardeningIntegrationTests.cs
@@ -498,7 +498,7 @@ internal sealed class BridgeSafetyCountingWormholeAdapter : IWormholeAdapter
 
     public Task<AZOAResult<WormholeTransferInitiation>> InitiateTransferAsync(
         string sourceChain, string targetChain, string tokenId,
-        string senderAddress, string recipientAddress, int amount,
+        string senderAddress, string recipientAddress, ulong amount,
         CancellationToken ct = default)
         => Task.FromResult(new AZOAResult<WormholeTransferInitiation>
         {
