+++ tests/AZOA.WebAPI.IntegrationTests/Gates/G2_IdempotencyTocTouTest.cs
@@ -446,7 +446,7 @@ internal sealed class CountingWormholeAdapter : IWormholeAdapter
 
     public Task<AZOAResult<WormholeTransferInitiation>> InitiateTransferAsync(
         string sourceChain, string targetChain, string tokenId,
-        string senderAddress, string recipientAddress, int amount,
+        string senderAddress, string recipientAddress, ulong amount,
         CancellationToken ct = default)
         => Task.FromResult(new AZOAResult<WormholeTransferInitiation>
         {
