+++ tests/AZOA.WebAPI.Tests/Services/Reconciliation/ReconciliationServiceTests.cs
@@ -800,21 +800,19 @@ public class ReconciliationServiceTests
         /// NEVER broadcasts/mutates on-chain.</summary>
         public void AssertNoOnChainMutation()
         {
-            // value-path-wiring H4: Mint/Burn/Transfer value amounts are ulong and
-            // carry a trailing SigningContext? param. Bridge methods (MintWrapped/
-            // BurnWrapped/LockForBridge) keep the int amount. The Times.Never
-            // assertion is unchanged — only the type-matchers track the new surface.
+            // value-path-wiring H4: all value amounts are ulong; direct chain calls
+            // also carry a trailing SigningContext? parameter.
             ProviderMock.Verify(p => p.MintAsync(
                 It.IsAny<string>(), It.IsAny<ulong>(), It.IsAny<string>(),
                 It.IsAny<string>(), It.IsAny<SigningContext?>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.MintWrappedAsync(
                 It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
-                It.IsAny<int>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
+                It.IsAny<ulong>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.BurnAsync(
                 It.IsAny<string>(), It.IsAny<ulong>(), It.IsAny<string>(),
                 It.IsAny<SigningContext?>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.BurnWrappedAsync(
-                It.IsAny<string>(), It.IsAny<int>(), It.IsAny<string>(),
+                It.IsAny<string>(), It.IsAny<ulong>(), It.IsAny<string>(),
                 It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.TransferAsync(
                 It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
@@ -826,7 +824,7 @@ public class ReconciliationServiceTests
                 It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
                 It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.LockForBridgeAsync(
-                It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int>(),
+                It.IsAny<string>(), It.IsAny<string>(), It.IsAny<ulong>(),
                 It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
             ProviderMock.Verify(p => p.DeployContractAsync(
                 It.IsAny<byte[]>(), It.IsAny<string>(),
